@@ -1,47 +1,66 @@
 {
-	"name": "codeigniter4/framework",
-	"type": "project",
-	"description": "The CodeIgniter framework v4",
-	"homepage": "https://codeigniter.com",
-	"license": "MIT",
-	"require": {
-		"php": "^7.3||^8.0",
-		"ext-curl": "*",
-		"ext-intl": "*",
-		"ext-json": "*",
-		"ext-mbstring": "*",
-		"kint-php/kint": "^3.3",
-		"laminas/laminas-escaper": "^2.6",
-		"psr/log": "^1.1"
-	},
-	"require-dev": {
-		"codeigniter4/codeigniter4-standard": "^1.0",
-		"fakerphp/faker": "^1.9",
-		"mikey179/vfsstream": "^1.6",
-		"phpunit/phpunit": "^9.1",
-		"predis/predis": "^1.1",
-		"squizlabs/php_codesniffer": "^3.3"
-	},
-	"suggest": {
-		"ext-fileinfo": "Improves mime type detection for files"
-	},
-	"autoload": {
-		"psr-4": {
-			"CodeIgniter\\": "system/"
-		},
-		"exclude-from-classmap": [
-			"**/Database/Migrations/**"
-		]
-	},
-	"scripts": {
-		"post-update-cmd": [
-			"CodeIgniter\\ComposerScripts::postUpdate"
-		],
-		"test": "phpunit"
-	},
-	"support": {
-		"forum": "http://forum.codeigniter.com/",
-		"source": "https://github.com/codeigniter4/CodeIgniter4",
-		"slack": "https://codeigniterchat.slack.com"
-	}
+    "name": "codeigniter4/framework",
+    "type": "project",
+    "description": "The CodeIgniter framework v4",
+    "homepage": "https://codeigniter.com",
+    "license": "MIT",
+    "require": {
+        "php": "^7.4 || ^8.0",
+        "ext-intl": "*",
+        "ext-json": "*",
+        "ext-mbstring": "*",
+        "laminas/laminas-escaper": "^2.9",
+        "psr/log": "^1.1"
+    },
+    "require-dev": {
+        "kint-php/kint": "^5.0.4",
+        "codeigniter/coding-standard": "^1.5",
+        "fakerphp/faker": "^1.9",
+        "friendsofphp/php-cs-fixer": "3.13.0",
+        "mikey179/vfsstream": "^1.6",
+        "nexusphp/cs-config": "^3.6",
+        "phpunit/phpunit": "^9.1",
+        "predis/predis": "^1.1 || ^2.0"
+    },
+    "suggest": {
+        "ext-curl": "If you use CURLRequest class",
+        "ext-imagick": "If you use Image class ImageMagickHandler",
+        "ext-gd": "If you use Image class GDHandler",
+        "ext-exif": "If you run Image class tests",
+        "ext-simplexml": "If you format XML",
+        "ext-mysqli": "If you use MySQL",
+        "ext-oci8": "If you use Oracle Database",
+        "ext-pgsql": "If you use PostgreSQL",
+        "ext-sqlsrv": "If you use SQL Server",
+        "ext-sqlite3": "If you use SQLite3",
+        "ext-memcache": "If you use Cache class MemcachedHandler with Memcache",
+        "ext-memcached": "If you use Cache class MemcachedHandler with Memcached",
+        "ext-redis": "If you use Cache class RedisHandler",
+        "ext-dom": "If you use TestResponse",
+        "ext-libxml": "If you use TestResponse",
+        "ext-xdebug": "If you use CIUnitTestCase::assertHeaderEmitted()",
+        "ext-fileinfo": "Improves mime type detection for files",
+        "ext-readline": "Improves CLI::input() usability"
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true
+    },
+    "autoload": {
+        "psr-4": {
+            "CodeIgniter\\": "system/"
+        },
+        "exclude-from-classmap": [
+            "**/Database/Migrations/**"
+        ]
+    },
+    "scripts": {
+        "test": "phpunit"
+    },
+    "support": {
+        "forum": "https://forum.codeigniter.com/",
+        "source": "https://github.com/codeigniter4/CodeIgniter4",
+        "slack": "https://codeigniterchat.slack.com"
+    }
 }
