@@ -1,70 +1,11 @@
 <?php
 
 namespace App\Controllers;
-use App\Models\Commonmodel;
-use CodeIgniter\API\ResponseTrait;
+
 class Home extends BaseController
 {
-	use ResponseTrait;
-	public function initController(\CodeIgniter\HTTP\RequestInterface $request, \CodeIgniter\HTTP\ResponseInterface $response, \Psr\Log\LoggerInterface $logger)
+    public function index()
     {
-        parent::initController($request, $response, $logger);
-        date_default_timezone_set('Asia/Karachi');
-        $this->Commonmodel = new Commonmodel();
+        return view('welcome_message');
     }
-	public function index()
-	{
-		$data['Employees']=$this->Commonmodel->rows_number('saimtech_employees');
-		$data['Departments']=$this->Commonmodel->rows_number('saimtech_departments');
-		$data['Documents']=$this->Commonmodel->rows_number('saimtech_document');
-		return view('dashboard/index',$data);
-	}
-
-	
-	public function myview()
-	{
-		return view('myView');
-	}
-	//---------------------Amount in words
-    public function AmountInWords(float $amount)
-	{
-	   $amount_after_decimal = round($amount - ($num = floor($amount)), 2) * 100;
-	   // Check if there is any number after decimal
-	   $amt_hundred = null;
-	   $count_length = strlen($num);
-	   $x = 0;
-	   $string = array();
-	   $change_words = array(0 => '', 1 => 'One', 2 => 'Two',
-	     3 => 'Three', 4 => 'Four', 5 => 'Five', 6 => 'Six',
-	     7 => 'Seven', 8 => 'Eight', 9 => 'Nine',
-	     10 => 'Ten', 11 => 'Eleven', 12 => 'Twelve',
-	     13 => 'Thirteen', 14 => 'Fourteen', 15 => 'Fifteen',
-	     16 => 'Sixteen', 17 => 'Seventeen', 18 => 'Eighteen',
-	     19 => 'Nineteen', 20 => 'Twenty', 30 => 'Thirty',
-	     40 => 'Forty', 50 => 'Fifty', 60 => 'Sixty',
-	     70 => 'Seventy', 80 => 'Eighty', 90 => 'Ninety');
-	    $here_digits = array('', 'Hundred','Thousand','Lakh', 'Crore');
-	    while( $x < $count_length ) {
-	      $get_divider = ($x == 2) ? 10 : 100;
-	      $amount = floor($num % $get_divider);
-	      $num = floor($num / $get_divider);
-	      $x += $get_divider == 10 ? 1 : 2;
-	      if ($amount) {
-	       $add_plural = (($counter = count($string)) && $amount > 9) ? 's' : null;
-	       $amt_hundred = ($counter == 1 && $string[0]) ? ' and ' : null;
-	       $string [] = ($amount < 21) ? $change_words[$amount].' '. $here_digits[$counter]. $add_plural.' 
-	       '.$amt_hundred:$change_words[floor($amount / 10) * 10].' '.$change_words[$amount % 10]. ' 
-	       '.$here_digits[$counter].$add_plural.' '.$amt_hundred;
-	        }
-	   else $string[] = null;
-	   }
-	   $implode_to_Rupees = implode('', array_reverse($string));
-	   $get_paise = ($amount_after_decimal > 0) ? "And " . ($change_words[$amount_after_decimal / 10] . " 
-	   " . $change_words[$amount_after_decimal % 10]) . ' Paise' : '';
-	   return ($implode_to_Rupees ? $implode_to_Rupees . 'Rupees ' : '') . $get_paise;
-	}
-	// $amt_words=855600.45; 
-	// $get_amount= AmountInWords($amt_words);
-	// echo $get_amount;
-	//---------------------end Amount in words
 }
