@@ -1,41 +1,10 @@
-    <meta charset="UTF-8">
-    <title>HRMS</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+<meta charset="utf-8">
+<title>Studio | Dashboard</title>
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<meta name="description" content="">
+<meta name="author" content="">
 
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/font/iconsmind-s/css/iconsminds.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/attendance.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/font/simple-line-icons/css/simple-line-icons.css" />
-    
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/vendor/bootstrap.min.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/vendor/bootstrap.rtl.only.min.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/vendor/fullcalendar.min.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/vendor/bootstrap-float-label.min.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/vendor/bootstrap-datepicker3.min.css" />
-
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/vendor/component-custom-switch.min.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/vendor/perfect-scrollbar.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/font/iconsmind-s/css/iconsminds.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/font/simple-line-icons/css/simple-line-icons.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/css/main.css" />
-  
-        <!--------------------Bootstrap Vue and Vue Load---------------------------------->
-    <!-- Load required  BootstrapVue CSS -->
-    <link type="text/css" rel="stylesheet" href="//unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue.min.css" />
-    <link rel="stylesheet" href="<?= base_url(); ?>/public/asset/fontawesome/css/all.min.css" />
-    <link rel="stylesheet" type="text/css" href="<?= base_url(); ?>/public/asset/datetimepicker/build/css/bootstrap-datetimepicker.min.css"/>
-    <link rel="stylesheet" href="bootstrap-datetimepicker.min.css">
-    <!-- Load polyfills to support older browsers -->
-    <script src="//polyfill.io/v3/polyfill.min.js?features=es2015%2CIntersectionObserver" crossorigin="anonymous"></script>
-    <!-- Load Vue followed by BootstrapVue -->
-    <script src="https://cdn.jsdelivr.net/npm/vue@2.6.12"></script> 
-    <script src="//unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue.min.js"></script>
-    <!-- Load the following for BootstrapVueIcons support -->
-    <script src="//unpkg.com/bootstrap-vue@latest/dist/bootstrap-vue-icons.min.js"></script>
-
-    <!-------------------------------------------------------------------------->
-
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/0.21.1/axios.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.29.1/moment.min.js"></script>
-    <script src="//cdn.jsdelivr.net/npm/sweetalert2@10"></script>
-    <script src="<?= base_url(); ?>/public/asset/js/vendor/jquery-3.3.1.min.js"></script>
-    <script src="<?= base_url(); ?>/public/asset/datetimepicker/build/js/bootstrap-datetimepicker.min.js"></script>
+<!-- ================== BEGIN core-css ================== -->
+<link href="<?= URL ?>/assets/css/vendor.min.css" rel="stylesheet">
+<link href="<?= URL ?>/assets/css/app.min.css" rel="stylesheet">
+<!-- ================== END core-css ================== -->
