@@ -1,384 +1,356 @@
-<div class="menu">
-        <div class="main-menu">
-            <div class="scroll">
-                <ul class="list-unstyled">
-                    <li>
-                        <a href="<?= base_url();?>/home">
-                            <i class="iconsminds-shop-4"></i>
-                            <span>Dashboards</span>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="#emplyees">
-                            <i class="iconsminds-business-man-woman"></i> Employees
-                        </a>
-                    </li>
-                    <?php if($_SESSION['user_power']=='Account' || $_SESSION['user_power']=='SE' ):?>
-                        <li>
-                            <a href="<?= base_url();?>/payroll1">
-                                <i class="iconsminds-library"></i> <span
-                                    class="d-inline-block">Payroll</span>
+    <!-- BEGIN #sidebar -->
+    <div id="sidebar" class="app-sidebar">
+        <!-- BEGIN scrollbar -->
+        <div class="app-sidebar-content" data-scrollbar="true" data-height="100%">
+            <!-- BEGIN menu -->
+            <div class="menu">
+                <div class="menu-header">Navigation</div>
+                <div class="menu-item active">
+                    <a href="index.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-laptop"></i></span>
+                        <span class="menu-text">Dashboard</span>
+                    </a>
+                </div>
+                <div class="menu-item">
+                    <a href="analytics.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-chart-pie"></i></span>
+                        <span class="menu-text">Analytics</span>
+                    </a>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="#" class="menu-link">
+                        <span class="menu-icon">
+                            <i class="fa fa-envelope"></i>
+                            <span class="menu-icon-label">6</span>
+                        </span>
+                        <span class="menu-text">Email</span>
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="email_inbox.html" class="menu-link">
+                                <span class="menu-text">Inbox</span>
                             </a>
-                        </li> 
-                    <?php endif ;?> 
-                    <li>
-                        <a href="#attendance">
-                            <i class="iconsminds-books"></i> Attendance
-                        </a>
-                    </li>
-                    <li>
-                        <a href="<?= base_url();?>/des">
-                            <i class="iconsminds-student-hat"></i> <span
-                                class="d-inline-block">Designations</span>
-                        </a>
-                    </li>
-
-                    <li>
-                        <a href="<?= base_url();?>/allowances">
-                            <i class="iconsminds-notepad"></i> <span
-                                class="d-inline-block">Allowances</span>
-                        </a>
-                    </li>
-
-                    <li>
-                        <a href="<?= base_url();?>/dep">
-                            <i class="iconsminds-museum"></i> <span
-                                class="d-inline-block">Departments</span>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="<?= base_url();?>/doc" target="_blank">
-                            <i class="iconsminds-library"></i> Docs
-                        </a>
-                    </li>
-                    <?php if($_SESSION['user_power']=='SE'):?>
-                        <li>
-                            <a href="<?= base_url();?>/users">
-                                <i class="iconsminds-students"></i> <span
-                                    class="d-inline-block">Users</span>
+                        </div>
+                        <div class="menu-item">
+                            <a href="email_compose.html" class="menu-link">
+                                <span class="menu-text">Compose</span>
                             </a>
-                        </li> 
-                    <?php endif ;?> 
-
-                </ul>
+                        </div>
+                        <div class="menu-item">
+                            <a href="email_detail.html" class="menu-link">
+                                <span class="menu-text">Detail</span>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-divider"></div>
+                <div class="menu-header">Components</div>
+                <div class="menu-item">
+                    <a href="widgets.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-qrcode"></i></span>
+                        <span class="menu-text">Widgets</span>
+                    </a>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="javascript:;" class="menu-link">
+                        <div class="menu-icon">
+                            <i class="fa fa-wallet"></i>
+                        </div>
+                        <div class="menu-text d-flex align-items-center">POS System</div> 
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="pos_customer_order.html" target="_blank" class="menu-link">
+                                <div class="menu-text">Customer Order</div>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="pos_kitchen_order.html" target="_blank" class="menu-link">
+                                <div class="menu-text">Kitchen Order</div>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="pos_counter_checkout.html" target="_blank" class="menu-link">
+                                <div class="menu-text">Counter Checkout</div>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="pos_table_booking.html" target="_blank" class="menu-link">
+                                <div class="menu-text">Table Booking</div>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="pos_menu_stock.html" target="_blank" class="menu-link">
+                                <div class="menu-text">Menu Stock</div>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="#" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-heart"></i></span>
+                        <span class="menu-text">UI Kits</span> 
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="ui_bootstrap.html" class="menu-link">
+                                <span class="menu-text">Bootstrap</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="ui_buttons.html" class="menu-link">
+                                <span class="menu-text">Buttons</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="ui_card.html" class="menu-link">
+                                <span class="menu-text">Card</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="ui_icons.html" class="menu-link">
+                                <span class="menu-text">Icons</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="ui_modal_notification.html" class="menu-link">
+                                <span class="menu-text">Modal & Notification</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="ui_typography.html" class="menu-link">
+                                <span class="menu-text">Typography</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="ui_tabs_accordions.html" class="menu-link">
+                                <span class="menu-text">Tabs & Accordions</span>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="#" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-file-alt"></i></span>
+                        <span class="menu-text">Forms</span> 
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="form_elements.html" class="menu-link">
+                                <span class="menu-text">Form Elements</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="form_plugins.html" class="menu-link">
+                                <span class="menu-text">Form Plugins</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="form_wizards.html" class="menu-link">
+                                <span class="menu-text">Wizards</span>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="#" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-table"></i></span>
+                        <span class="menu-text">Tables</span>
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="table_elements.html" class="menu-link">
+                                <span class="menu-text">Table Elements</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="table_plugins.html" class="menu-link">
+                                <span class="menu-text">Table Plugins</span>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="#" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-chart-bar"></i></span>
+                        <span class="menu-text">Charts</span>
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="chart_js.html" class="menu-link">
+                                <span class="menu-text">Chart.js</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="chart_apex.html" class="menu-link">
+                                <span class="menu-text">Apexcharts.js</span>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-item">
+                    <a href="map.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-map-marker-alt"></i></span>
+                        <span class="menu-text">Map</span>
+                    </a>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="#" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-code-branch"></i></span>
+                        <span class="menu-text">Layout</span>
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="layout_starter.html" class="menu-link">
+                                <span class="menu-text">Starter Page</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="layout_fixed_footer.html" class="menu-link">
+                                <span class="menu-text">Fixed Footer</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="layout_full_height.html" class="menu-link">
+                                <span class="menu-text">Full Height</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="layout_full_width.html" class="menu-link">
+                                <span class="menu-text">Full Width</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="layout_boxed_layout.html" class="menu-link">
+                                <span class="menu-text">Boxed Layout</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="layout_minified_sidebar.html" class="menu-link">
+                                <span class="menu-text">Minified Sidebar</span>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-item has-sub">
+                    <a href="#" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-globe"></i></span>
+                        <span class="menu-text">Pages</span>
+                        <span class="menu-caret"><b class="caret"></b></span>
+                    </a>
+                    <div class="menu-submenu">
+                        <div class="menu-item">
+                            <a href="page_scrum_board.html" class="menu-link">
+                                <span class="menu-text">Scrum Board</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_products.html" class="menu-link">
+                                <span class="menu-text">Products</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_product_details.html" class="menu-link">
+                                <span class="menu-text">Product Details</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_orders.html" class="menu-link">
+                                <span class="menu-text">Orders</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_order_details.html" class="menu-link">
+                                <span class="menu-text">Order Details</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_gallery.html" class="menu-link">
+                                <span class="menu-text">Gallery</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_search_results.html" class="menu-link">
+                                <span class="menu-text">Search Results</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_coming_soon.html" class="menu-link">
+                                <span class="menu-text">Coming Soon Page</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_404_error.html" class="menu-link">
+                                <span class="menu-text">404 Error Page</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_login.html" class="menu-link">
+                                <span class="menu-text">Login</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_register.html" class="menu-link">
+                                <span class="menu-text">Register</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_messenger.html" class="menu-link">
+                                <span class="menu-text">Messenger</span>
+                            </a>
+                        </div>
+                        <div class="menu-item">
+                            <a href="page_data_management.html" class="menu-link">
+                                <span class="menu-text">Data Management</span>
+                            </a>
+                        </div>
+                    </div>
+                </div>
+                <div class="menu-divider"></div>
+                <div class="menu-header">Users</div>
+                <div class="menu-item">
+                    <a href="profile.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-user-circle"></i></span>
+                        <span class="menu-text">Profile</span>
+                    </a>
+                </div>
+                <div class="menu-item">
+                    <a href="calendar.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-calendar"></i></span>
+                        <span class="menu-text">Calendar</span>
+                    </a>
+                </div>
+                <div class="menu-item">
+                    <a href="settings.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-cog"></i></span>
+                        <span class="menu-text">Settings</span>
+                    </a>
+                </div>
+                <div class="menu-item">
+                    <a href="helper.html" class="menu-link">
+                        <span class="menu-icon"><i class="fa fa-question-circle"></i></span>
+                        <span class="menu-text">Helper</span>
+                    </a>
+                </div>
+                <div class="p-3 px-4 mt-auto hide-on-minified">
+                    <a href="https://seantheme.com/studio/documentation/index.html" class="btn btn-secondary d-block w-100 fw-600 rounded-pill">
+                        <i class="fa fa-code-branch me-1 ms-n1 opacity-5"></i> Documentation
+                    </a>
+                </div>
             </div>
+            <!-- END menu -->
         </div>
-
-        <div class="sub-menu">
-            <div class="scroll">
-                <ul class="list-unstyled" data-link="emplyees" id="emplyees">
-                    <li>
-                        <a href="#" data-toggle="collapse" data-target="#collapseAuthorization" aria-expanded="true"
-                            aria-controls="collapseAuthorization" class="rotate-arrow-icon opacity-50">
-                            <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Pages</span>
-                        </a>
-                        <div id="collapseAuthorization" class="collapse show">
-                            <ul class="list-unstyled inner-level-menu">
-                                <li>
-                                    <a href="<?= base_url();?>/employee">
-                                        <i class="iconsminds-business-man-woman"></i> <span
-                                            class="d-inline-block">Employees</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="<?= base_url();?>/addemployee">
-                                        <i class="iconsminds-add-user"></i> <span
-                                            class="d-inline-block">Emp Registration</span>
-                                    </a>
-                                </li>
-                                  
-                                 
-
-                                <!-- <li>
-                                    <a href="<?= base_url();?>/LeaveTypes">
-                                        <i class="iconsminds-library"></i> <span
-                                            class="d-inline-block">Leave Types</span>
-                                    </a>
-                                </li> -->
-<!--                                 <li>
-                                    <a href="<?= base_url();?>/Payroll1">
-                                        <i class="iconsminds-library"></i> <span
-                                            class="d-inline-block">Payroll</span>
-                                    </a>
-                                </li> -->                           
-                            </ul>
-                        </div>
-                    </li>
-                   
-                </ul>
-                <ul class="list-unstyled" data-link="attendance">
-                    <li>
-                        <a href="<?= base_url();?>/Attendance">
-                            <i class="iconsminds-students"></i> <span
-                                class="d-inline-block">Attendance</span>
-                        </a>
-                    </li>
-                    <li>
-                        <a href="<?= base_url();?>/Attendance/attendancereport">
-                            <i class="iconsminds-library"></i> <span
-                                class="d-inline-block">Attendance Report</span>
-                        </a>
-                    </li>                 
-                </ul>
-                <ul class="list-unstyled" data-link="ui">
-                    <li>
-                        <a href="#" data-toggle="collapse" data-target="#collapseForms" aria-expanded="true"
-                            aria-controls="collapseForms" class="rotate-arrow-icon opacity-50">
-                            <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Forms</span>
-                        </a>
-                        <div id="collapseForms" class="collapse show">
-                            <ul class="list-unstyled inner-level-menu">
-                                <li>
-                                    <a href="Ui.Forms.Components.html">
-                                        <i class="simple-icon-event"></i> <span class="d-inline-block">Components</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Forms.Layouts.html">
-                                        <i class="simple-icon-doc"></i> <span class="d-inline-block">Layouts</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Forms.Validation.html">
-                                        <i class="simple-icon-check"></i> <span class="d-inline-block">Validation</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Forms.Wizard.html">
-                                        <i class="simple-icon-magic-wand"></i> <span
-                                            class="d-inline-block">Wizard</span>
-                                    </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </li>
-                    <li>
-                        <a href="#" data-toggle="collapse" data-target="#collapseDataTables" aria-expanded="true"
-                            aria-controls="collapseDataTables" class="rotate-arrow-icon opacity-50">
-                            <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Datatables</span>
-                        </a>
-                        <div id="collapseDataTables" class="collapse show">
-                            <ul class="list-unstyled inner-level-menu">
-                                <li>
-                                    <a href="Ui.Datatables.Rows.html">
-                                        <i class="simple-icon-screen-desktop"></i> <span class="d-inline-block">Full
-                                            Page UI</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Datatables.Scroll.html">
-                                        <i class="simple-icon-mouse"></i> <span class="d-inline-block">Scrollable</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Datatables.Pagination.html">
-                                        <i class="simple-icon-notebook"></i> <span
-                                            class="d-inline-block">Pagination</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Datatables.Default.html">
-                                        <i class="simple-icon-grid"></i> <span class="d-inline-block">Default</span>
-                                    </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </li>
-                    <li>
-                        <a href="#" data-toggle="collapse" data-target="#collapseComponents" aria-expanded="true"
-                            aria-controls="collapseComponents" class="rotate-arrow-icon opacity-50">
-                            <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Components</span>
-                        </a>
-                        <div id="collapseComponents" class="collapse show">
-                            <ul class="list-unstyled inner-level-menu">
-                                <li>
-                                    <a href="Ui.Components.Alerts.html">
-                                        <i class="simple-icon-bell"></i> <span class="d-inline-block">Alerts</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Badges.html">
-                                        <i class="simple-icon-badge"></i> <span class="d-inline-block">Badges</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Buttons.html">
-                                        <i class="simple-icon-control-play"></i> <span
-                                            class="d-inline-block">Buttons</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Cards.html">
-                                        <i class="simple-icon-layers"></i> <span class="d-inline-block">Cards</span>
-                                    </a>
-                                </li>
-
-                                <li>
-                                    <a href="Ui.Components.Carousel.html">
-                                        <i class="simple-icon-picture"></i> <span class="d-inline-block">Carousel</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Charts.html">
-                                        <i class="simple-icon-chart"></i> <span class="d-inline-block">Charts</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Collapse.html">
-                                        <i class="simple-icon-arrow-up"></i> <span
-                                            class="d-inline-block">Collapse</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Dropdowns.html">
-                                        <i class="simple-icon-arrow-down"></i> <span
-                                            class="d-inline-block">Dropdowns</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Editors.html">
-                                        <i class="simple-icon-book-open"></i> <span
-                                            class="d-inline-block">Editors</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Icons.html">
-                                        <i class="simple-icon-star"></i> <span class="d-inline-block">Icons</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.InputGroups.html">
-                                        <i class="simple-icon-note"></i> <span class="d-inline-block">Input
-                                            Groups</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Jumbotron.html">
-                                        <i class="simple-icon-screen-desktop"></i> <span
-                                            class="d-inline-block">Jumbotron</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Modal.html">
-                                        <i class="simple-icon-docs"></i> <span class="d-inline-block">Modal</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Navigation.html">
-                                        <i class="simple-icon-cursor"></i> <span
-                                            class="d-inline-block">Navigation</span>
-                                    </a>
-                                </li>
-
-                                <li>
-                                    <a href="Ui.Components.PopoverandTooltip.html">
-                                        <i class="simple-icon-pin"></i> <span class="d-inline-block">Popover &
-                                            Tooltip</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Sortable.html">
-                                        <i class="simple-icon-shuffle"></i> <span class="d-inline-block">Sortable</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Ui.Components.Tables.html">
-                                        <i class="simple-icon-grid"></i> <span class="d-inline-block">Tables</span>
-                                    </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </li>
-
-                </ul>
-
-                <ul class="list-unstyled" data-link="menu" id="menuTypes">
-                    <li>
-                        <a href="#" data-toggle="collapse" data-target="#collapseMenuTypes" aria-expanded="true"
-                            aria-controls="collapseMenuTypes" class="rotate-arrow-icon">
-                            <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Menu Types</span>
-                        </a>
-                        <div id="collapseMenuTypes" class="collapse show" data-parent="#menuTypes">
-                            <ul class="list-unstyled inner-level-menu">
-                                <li>
-                                    <a href="Menu.Default.html">
-                                        <i class="simple-icon-control-pause"></i> <span
-                                            class="d-inline-block">Default</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Menu.Subhidden.html">
-                                        <i class="simple-icon-arrow-left mi-subhidden"></i> <span
-                                            class="d-inline-block">Subhidden</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Menu.Hidden.html">
-                                        <i class="simple-icon-control-start mi-hidden"></i> <span
-                                            class="d-inline-block">Hidden</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="Menu.Mainhidden.html">
-                                        <i class="simple-icon-control-rewind mi-hidden"></i> <span
-                                            class="d-inline-block">Mainhidden</span>
-                                    </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </li>
-                    <li>
-                        <a href="#" data-toggle="collapse" data-target="#collapseMenuLevel" aria-expanded="true"
-                            aria-controls="collapseMenuLevel" class="rotate-arrow-icon collapsed">
-                            <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Menu Levels</span>
-                        </a>
-                        <div id="collapseMenuLevel" class="collapse" data-parent="#menuTypes">
-                            <ul class="list-unstyled inner-level-menu">
-                                <li>
-                                    <a href="#">
-                                        <i class="simple-icon-layers"></i> <span class="d-inline-block">Sub
-                                            Level</span>
-                                    </a>
-                                </li>
-                                <li>
-                                    <a href="#" data-toggle="collapse" data-target="#collapseMenuLevel2"
-                                        aria-expanded="true" aria-controls="collapseMenuLevel2"
-                                        class="rotate-arrow-icon collapsed">
-                                        <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Another
-                                            Level</span>
-                                    </a>
-                                    <div id="collapseMenuLevel2" class="collapse">
-                                        <ul class="list-unstyled inner-level-menu">
-                                            <li>
-                                                <a href="#">
-                                                    <i class="simple-icon-layers"></i> <span class="d-inline-block">Sub
-                                                        Level</span>
-                                                </a>
-                                            </li>
-                                        </ul>
-                                    </div>
-                                </li>
-                            </ul>
-                        </div>
-                    </li>
-                    <li>
-                        <a href="#" data-toggle="collapse" data-target="#collapseMenuDetached" aria-expanded="true"
-                            aria-controls="collapseMenuDetached" class="rotate-arrow-icon collapsed">
-                            <i class="simple-icon-arrow-down"></i> <span class="d-inline-block">Detached</span>
-                        </a>
-                        <div id="collapseMenuDetached" class="collapse">
-                            <ul class="list-unstyled inner-level-menu">
-                                <li>
-                                    <a href="#">
-                                        <i class="simple-icon-layers"></i> <span class="d-inline-block">Sub
-                                            Level</span>
-                                    </a>
-                                </li>
-                            </ul>
-                        </div>
-                    </li>
-                </ul>
-                
-            </div>
-        </div>
+        <!-- END scrollbar -->
+        
+        <!-- BEGIN mobile-sidebar-backdrop -->
+        <button class="app-sidebar-mobile-backdrop" data-dismiss="sidebar-mobile"></button>
+        <!-- END mobile-sidebar-backdrop -->
     </div>
+    <!-- END #sidebar -->
